Update returnofdiagnol.py
Your code calculates the number of columns by only considering the first row of the matrix. This is fine if you have the assumption that all rows have the same number of columns. However, to avoid this assumption, we can use numpy.shape() I have implemented that change for you